+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to
 - [#5165](https://github.com/firecracker-microvm/firecracker/pull/5165): Changed
   Firecracker snapshot feature from developer preview to generally available.
   Incremental snapshots remain in developer preview.
+- [#5282](https://github.com/firecracker-microvm/firecracker/pull/5282): Update
+  jailer executable requrements. Now jailer does require for an executable to
+  have exec permissions set, but does not require it's name to contain
+  `firecracker`.
 
 ### Deprecated